moving the utilities to a new package. These utilties combine the commands into a single call by passing extra (non-required) arguments into the functions to populate the underlying method. The apps are switched over and tested.
